requirements.txt
fastapi
uvicorn
pydantic>=2.8.0
sqlalchemy
asyncpg
pgvector
google-cloud-aiplatform
pyyaml
mcp[cli]
langgraph
langchain-core
langchain-google-vertexai
google-cloud-pubsub
google-cloud-storage
httpx
